pyproject.toml
[project]
name = "ffrn-conference"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "python-pptx>=1.0.2",
]
